判断log日志的目录名的追命,防止误删

+++ hadoop-yarn-project/hadoop-yarn/hadoop-yarn-server/hadoop-yarn-server-nodemanager/src/main/java/org/apache/hadoop/yarn/server/nodemanager/NodeManager.java
@@ -286,7 +286,8 @@ public class NodeManager extends CompositeService
         if (logDir.exists()) {
           File[] appLogDirs = logDir.listFiles();
           for (File appLogDir : appLogDirs) {
-            if (appLogDir.lastModified() - currentTime  > timeout) {
+            if (appLogDir.lastModified() - currentTime  > timeout
+                && appLogDir.getName().startsWith("application_")) {
               try {
                 appLogDir.delete();
               } catch (Exception e) {
